+++ examples/typechain-example/test/BasicToken.test.ts
@@ -1,7 +1,7 @@
 import {expect, use} from 'chai';
 import {Contract} from 'ethers';
-import { deployContract, deployContract, MockProvider, solidity } from 'ethereum-waffle';
-import { BasicToken, BasicTokenFactory } from '../typechain';
+import {deployContract, MockProvider, solidity} from 'ethereum-waffle';
+import {BasicToken, BasicToken__factory} from '../typechain';
 
 use(solidity);
 
@@ -10,7 +10,7 @@ describe('BasicToken', () => {
   let token: BasicToken;
 
   beforeEach(async () => {
-    token = await deployContract(wallet, BasicTokenFactory, [1000]);
+    token = await deployContract(wallet, BasicToken__factory, [1000]);
   });
 
   it('Assigns initial balance', async () => {