+++ waffle-jest/src/matchers/toBeReverted.ts
@@ -1,13 +1,25 @@
 export async function toBeReverted(promise: Promise<any>) {
   try {
-    await promise;
+    const tx = await promise;
+    if ('wait' in tx) {
+      // Sending the transaction succeeded, but we wait to see if it will revert on-chain.
+      try {
+        await tx.wait();
+      } catch (e) {
+        return {
+          pass: true,
+          message: () => 'Expected transaction to be reverted'
+        };
+      }
+    }
+
     return {
       pass: false,
       message: () => 'Expected transaction to be reverted'
     };
   } catch (error) {
     const message =
-      error instanceof Object && 'message' in error ? error.message : JSON.stringify(error);
+      error instanceof Object && 'message' in error ? (error as any).message : JSON.stringify(error);
 
     const isReverted = message.search('revert') >= 0;
     const isThrown = message.search('invalid opcode') >= 0;